% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_dictionary_language.R
\name{choose_dictionary_language}
\alias{choose_dictionary_language}
\title{Extract translation strings from translation data.frame}
\usage{
choose_dictionary_language(data, language)
}
\arguments{
\item{data}{data.frame containing translation data, with columns \code{id_data},
\code{id_column}, \code{key}, and one column for each language (lowercase ISO-639-1
code). See examples for more detail.}

\item{language}{ISO-639-1 code of the language to extract from the
translation data.}
}
\value{
data.frame with tranlsated strings for the given language in the
\code{translate_value} column.
}
\description{
Extract the translation strings from a translation data.frame for a given
language.
}
\examples{
translation_data <- read.csv(
  file = system.file(
    "extdata", "translation", "dataframe_labels.csv",
    package = "pacta.portfolio.report"
  )
)
de_dictionary <- choose_dictionary_language(translation_data, "de")
}
