% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactive_report.R
\name{create_interactive_report}
\alias{create_interactive_report}
\title{Create interactive report}
\usage{
create_interactive_report(
  template_dir = NULL,
  output_dir = NULL,
  survey_dir = "",
  real_estate_dir = "",
  language_select = "EN",
  investor_name = NULL,
  portfolio_name = NULL,
  peer_group = NULL,
  start_year = NULL,
  select_scenario,
  select_scenario_other,
  portfolio_allocation_method = NULL,
  scenario_geography = NULL,
  pacta_sectors = c("Power", "Automotive", "Oil&Gas", "Coal", "Steel", "Cement",
    "Aviation"),
  green_techs = c("RenewablesCap", "HydroCap", "NuclearCap", "Hybrid", "Electric",
    "FuelCell", "Hybrid_HDV", "Electric_HDV", "FuelCell_HDV", "Electric Arc Furnace"),
  tech_roadmap_sectors = c("Automotive", "Power", "Oil&Gas", "Coal"),
  pacta_sectors_not_analysed = c("Aviation", "Cement", "Steel"),
  audit_file = NULL,
  emissions = NULL,
  portfolio_overview = NULL,
  equity_results_portfolio = NULL,
  bonds_results_portfolio = NULL,
  equity_results_company = NULL,
  bonds_results_company = NULL,
  equity_results_map = NULL,
  bonds_results_map = NULL,
  indices_equity_results_portfolio = NULL,
  indices_bonds_results_portfolio = NULL,
  peers_equity_results_portfolio = NULL,
  peers_bonds_results_portfolio = NULL,
  peers_equity_results_user = NULL,
  peers_bonds_results_user = NULL,
  display_currency = "USD",
  currency_exchange_value = 1,
  dataframe_translations = NULL,
  js_translations = NULL,
  header_dictionary = NULL,
  sector_order = NULL,
  configs = NULL
)
}
\arguments{
\item{template_dir}{a parameter}

\item{output_dir}{a parameter}

\item{survey_dir}{a parameter}

\item{real_estate_dir}{a parameter}

\item{language_select}{a parameter}

\item{investor_name}{a parameter}

\item{portfolio_name}{a parameter}

\item{peer_group}{a parameter}

\item{start_year}{a parameter}

\item{select_scenario}{a parameter}

\item{select_scenario_other}{a parameter}

\item{portfolio_allocation_method}{a parameter}

\item{scenario_geography}{a parameter}

\item{pacta_sectors}{a parameter}

\item{green_techs}{a parameter}

\item{tech_roadmap_sectors}{a parameter}

\item{pacta_sectors_not_analysed}{a parameter}

\item{audit_file}{a parameter}

\item{emissions}{a parameter}

\item{portfolio_overview}{a parameter}

\item{equity_results_portfolio}{a parameter}

\item{bonds_results_portfolio}{a parameter}

\item{equity_results_company}{a parameter}

\item{bonds_results_company}{a parameter}

\item{equity_results_map}{a parameter}

\item{bonds_results_map}{a parameter}

\item{indices_equity_results_portfolio}{a parameter}

\item{indices_bonds_results_portfolio}{a parameter}

\item{peers_equity_results_portfolio}{a parameter}

\item{peers_bonds_results_portfolio}{a parameter}

\item{peers_equity_results_user}{a parameter}

\item{peers_bonds_results_user}{a parameter}

\item{display_currency}{a parameter}

\item{currency_exchange_value}{a parameter}

\item{dataframe_translations}{a parameter}

\item{js_translations}{a parameter}

\item{header_dictionary}{a parameter}

\item{sector_order}{a parameter}

\item{configs}{a parameter}
}
\value{
only used for side-effect
}
\description{
This function creates an interactive report
}
