% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_template_path.R
\name{get_report_template_path}
\alias{get_report_template_path}
\title{Get the path to the template based on report name and language}
\usage{
get_report_template_path(
  project_report_name = "general",
  language_select = "EN"
)
}
\arguments{
\item{project_report_name}{a single string specifying the name of the report, e.g. "GENERAL"}

\item{language_select}{a single string specifying the name of the language code, e.g. "EN"}
}
\value{
A string containing the path to the template directory
}
\description{
This function returns the path to the template based on the report name and
the language code. The templates are installed with the package. If a
template is not found that corresponds to the specified report name and
language code, a default template is used instead.
}
