% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_report_template.R
\name{select_report_template}
\alias{select_report_template}
\title{select correct report template}
\usage{
select_report_template(project_report_name = NULL, language_select = "EN")
}
\arguments{
\item{project_report_name}{a parameter}

\item{language_select}{a parameter}
}
\value{
A string containing the name of the template
}
\description{
This function returns the name of the template
}
